requirements.txt
# Root requirements.txt — used by Railway / Render for deployment.
# IMPORTANT: CPU-only torch keeps the image under 4GB.

--extra-index-url https://download.pytorch.org/whl/cpu
torch

# Server
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.12

# LangGraph / LangChain
langgraph==0.2.50
langchain==0.3.7
langchain-groq==0.2.1
langchain-community==0.3.7
langchain-openai==0.2.6

# LLM SDKs
openai==1.54.0
groq==0.11.0

# Vector Store / RAG
faiss-cpu==1.9.0
pypdf==4.3.1
tiktoken==0.8.0
langchain-huggingface>=0.1.0
langchain-text-splitters>=0.3.0
sentence-transformers>=3.0.0

# Schema & Config
pydantic==2.9.2
python-dotenv==1.0.1

# Utilities
httpx==0.27.2
aiofiles==24.1.0
